Added the a new Activity for setting the MovieInfo sort order
Created SettingsActivity that extends PreferenceActivity and added the
pref_general.xml resource that defines the settings screen.

Added a ListPreference in pref_general.xml that handles the selection
for the movie sort order.

Added the Intent call for SettingsActivity when the settings option has
been selected in ActivityMain.

Updated the api request to use the sortOrder preference.